imx8m/soc.c: deleting phandles pointing to deleted cpus
For i.MX8M Mini with less than 4 CPU cores the code patches the device
tree to reflect that and deletes the nodes representing the not existing
cores.

The deleted nodes are referenced by phandles in the following two properties:
/thermal-zones/cpu-thermal/cooling-maps/map0/cooling-device
/pmu/interrupt-affinity

Modify these properties to no longer contain the phandles to no longer
existing nodes.

Fixes the thermal subsystem which otherwise bails out with:

[    0.102895] OF: /thermal-zones/cpu-thermal/cooling-maps/map0: could not find phandle
[    0.102922] thermal_sys: failed to build thermal zone cpu-thermal: -22
[    1.377046] i.mx8mm_thermal 30260000.tmu: failed to register thermal zone sensor[0]: 0
[    3.013718] i.mx8mm_thermal 30260000.tmu: failed to register thermal zone sensor[0]: 0
     ...

Related-to: ELB-3292
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>