ARM: dts: verdin-imx8mp: fix mac address passing
Verdin iMX8M Plus on-module Ethernet device is eqos, while the optional
secondary Ethernet is fec, update the Ethernet aliases in the .dts file
so that ethaddr is set to eqos and eth1addr to fec.

Without this change, the Ethernet interfaces may have different MAC
addresses between Linux and U-Boot.

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>