configs: verdin-imx8mp: Disable cmd env erase
Erasing environment on emmc using the `env erase` command will
erase surrounding data in case the env location is not aligned to the
erase group size.

This is the intended behavior, to be able to use the functionality
the environment should be aligned/multiple of the erase size.

This is not the case on verdin-imx8mp, when the environment
is erased also the Toradex config block is lost, to avoid issues
just disable the command in the defconfig.

Link: https://lore.kernel.org/u-boot/20211129152530.GI24579@bill-the-cat/
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>